ENFIN. Saloperie de bash.

+++ template/tests/do_exec.sh
@@ -7,19 +7,25 @@ echo "exec/"
 
 for f in exec/*.scala; do
     max=`expr $max + 1`;
-    echo $f
+    echo "* ""$f"
     rm -f out
     ../pscala $f && gcc ./lol.s;
     ./a.out > out;
-    filename="${f%.*}";
-    fout="$filename"".out";
-    rm -f truc;
-    `diff out $fout` > truc;
-    if grep -q '<' truc; then
-        echo "Failed: The compilation should have succeded";
+    if [ $? -eq 0 ]
+    then
+        filename="${f%.*}";
+        fout="$filename"".out";
+        rm -f truc;
+        diff out $fout > truc;
+        if grep -q '<' truc
+        then
+            echo 'Failed: Wrong output, you did shit !'
+        else
+            echo 'Success ! <3'
+            score=`expr $score + 1`
+        fi
     else
-        echo "Succès !";
-        score=`expr $score + 1`
+        echo 'Failed: Unexpected runtime error'
     fi
 done
 echo
